Replace libuuid with the Provide that works correctly on both Fedora and openSUSE.
The package name on Fedora is libuuid, but the package name on openSUSE is libuuid1. RPM packages don't support a dependency on (package A OR package B) (see https://stackoverflow.com/q/2139621/1915961), they require you to use a common virtual provide. Inspecting the libuuid/libuuid1 packages on x86-64 and ARM64 for both Fedora and openSUSE, the common virtual provide is libuuid.so.1()(64bit).